remove escape character from pexpect.spawn.interact
The default escape character for pexpect.spaw is ctrl-]
which conflicts with a vim binding. This isn't an escape
sequence that is mirrored in virtualenv, so there doesn't
seem to be a good reason to keep it. This will make `pipenv
shell` more inline with the standard `virtualenv workon` command.