Update pip's pkg_resources from Setuptools 34.2.0.
This is basically a port of pip's commit
eaccb88674beff75bb98a1d1e2d53a26a9c63890, which updates the vendored
pkg_resources, and fixes a crash which, apparently, seems to happen when
trying to import setuptools/pkg_resources in an environment where a
namespace package is installed, but where there is more than one
directory in the path of the namespace package.

This crash is reproducible by running any pipenv command, including
`pipenv --version` in an empty directory.

See the upstream bug here[1] for further details.

Fixes #552

[1] https://github.com/pypa/setuptools/issues/885