Fix __delitem__ in prettytoml's table element
Two bugs involved here.

First of all, __delitem__ uses _find_following_newline to find where the
line to delete ends, which in turns looks for the next closest newline
element. But if the previous line ends with a comment element (which
implies a newline), it will also be deleted. This would be a problem if
the comment is inline.

foo=1  # comment
bar=2

In this minimal example, the elements in play are

AtomicElement('foo')
PunctuationElement('=')
AtomicElement(1)
WhitespaceElement('  ')
CommentElement('# comment')
AtomicElement('bar')
PunctuationElement('=')
AtomicElement(2)

When deleting "foo", it starts from AtomicElement('foo') and looks for
the next newline, but does not consider CommentElement, ending up
deleting all elements. This fix replaces _find_following_newline with
_find_following_line_terminator, which does consider both newline and
comment elements.

Also there is a logic error in _find_following_line_terminator that
prevents it from even working. The two -inf/inf clauses should serve as
guards when either or both indexes are infinity, but they were
flip-flopped and not working at all. This change doesn't seem to affect
existing tests (at least not breaking them more than they already were).

Fix #1535.