Ask py.exe for its opinion when choosing a Python
Hopefully a viable alternative to #1620. In addition to looking for
executables in PATH, also asks py.exe (if there is one) for its opinion
when choosing an executable. Windows users would be able to use the
--two, --three, and --python=<VERSION> syntax as others do; this will be
passed on to py.exe and ask if it can find any matching executable.

Non-Windows users will not be affected because there wouldn't be a py.exe
to start with.