Improve resolver path-patching
Add the parent site-packages directory to sys.path *after* built-in
ones, and only when it does not already exist.

This should vastly improve the resolver's robustness against hostile
Python environments, such as #1753. It might help with issues like #693
as well, but I'm not sure (and not interested in figuring out either).