Remove extraneous is_vcs calls
1. is_vcs requires (req.path or req.uri) to be truthy.
2. (req.path or req.uri) would be true in first if statement

Thus if you have:

```
if (req.uri or req.path or ...) and not req.vcs
elif req.vcs or is_vcs(req.path)
```

if `is_vcs(req.path)` is Truthy, then the first if clause must be
truthy, so it's unnecessary.

Also add a quick test to confirm that originating purpose for that check
is now handled otherwise.