Add test of behaviour in issue #2229
requirements.txt files (and by extension pipfreeze) can contain comment
lines, but the stdout of pip freeze is used in a number of places with
the assumption that every line contains a dependency declaration. This
change adds an integration test that could trigger one of these cases
that was the cause of the behaviour seen in issue #2229.