Allow virtual env creation in dir with leading dash
If the name given to a virtualenv starts with a dash, it can be
misinterpreted as an argument when passed to external commands,
such as when pew is used create a new virtual environment.

Since pipenv has not been able to create a virtual environment in
a directory whose name begins with a dash, there are no deployments
to break with this change.

Fixes #439