Update deployment section of advanced docs
Clarify that `--deploy` is meant for enforcement and intentionally
failing a build.
Mention `pipenv sync` in the same section to try to establish it more
clearly as the most obvious and normal way of doing installation from
``Pipfile.lock`` as part of a reproducible build process.

Explicitly recommend against `pipenv install --ignore-pipfile` and
suggest that it may be deprecated.