Clarify `pipenv install` documentation
`if no packages are given` sounds a bit better than
`if none is given` when referring to `provided packages`.

Also mention that `pipenv install` with no provided packages
installs all packages from Pipfile. It is not obvious for
people new to pipenv whether pipenv is using Pipfile or
Pipfile.lock by default.