Incorrect ref in VCS example
Small changes, but a couple of issues in the `A Note About VCS Dependencies` section of the docs.

Running the example `$ pipenv install -e git+https://github.com/requests/requests.git@v2.19#egg=requests` raises the error: `Did not find branch or tag 'v2.19', assuming revision or ref.` I think it should be 'v2.19.1' which is just a typo as 'v2.19.1' is used in the subsequent lines of the example.

The example Pipfile in the  section excludes the `v` in the version number written to the `ref = ` part. I found it a little confusing that it would strip the 'v' from the tag, but then on testing it I found that it doesn't do that:

```
[packages]
requests = {editable = true, ref = "v2.20.1", git = "https://github.com/requests/requests.git"}
```

Also, while I was here I figured I'd update the example to the most recent requests release.

Thanks.