Fix a TOCTOU issue in mkdir_p
When running multiple pipenv processes simultaneously, they may both
attempt to use 'mkdir_p' at the same time, which can cause one of the
pipenv processes to crash when it tries to create a directory that
already exists.

Instead of crashing outright, this commit instead opts to continue
onward if the 'os.mkdir' call failed with EEXIST, which is fine since
we're trying to create the directory anyway after ascertaining that it
didn't exist before.