Update basics.rst
Changed the scheme in VCS 

because with a colon instead of a slash cloning of the repository fails. This is due to the failing of git clone when ssh:// is used, a colon is considered the delimiter for the host's port and not the username. Please see https://github.com/pypa/pipenv/issues/3410 for this.