[#3722] Exclude shared packages for all-dev opt
Filters list of packages to uninstall for shared dependencies when
uninstall is used with the --all-dev option. This prevents shared
dependencies from being uninstalled or removed from the default section
of the lockfile.

Adds a test to ensure that shared dependencies are not uninstalled and
not removed from the pipfile or lockfile.