core.pip_install: Quote tokens when logging pip commands
Previously, pip commands were printed as lists of tokens. a82bbb7b amended the
logging to join the tokens with strings. However, this will lead to invalid
commands if any token contains a space or some other special character that
needs to be escaped for standard shells. In order to improve the logging, this
commit shell-quotes each token before joining the results with spaces.