Remove vistir.misc.dedup
This function is a nice to have alias for one line of code.
Also, it uses OrderedDict, which is no longer needed. `dict`s after
Python3.6 are guaranteed to be ordered (bonus: dict is faster than
ordered dict).

If we really want, we can also import it from zipp or python finder.