fix(environment): update python executable lookup
This fix resolves a faulty path on MSYS2 MinGW CPython, where the bin/ segment
is missing altogether. This is required by the uninstall routine. It also
simplifies the lookup on all other platforms.

As far as I comprehended, the condition in line 255 would never be met, as the
joinpath('python') in line 254 guarantees that py will never become a
false-equivalent value, since its definition is in the else statement, hence
catches the unmatched condition in line 251.

I've also inverted the condition of checking for self._python, so that an
intermediate assignment to `py` is no longer required.