Correct check/scan command documentation
In commit 56d45354e12fd9a3b8cd13dc64ac8566225a29a8 the new `--scan`
option for the `check` command was introduced, and a deprecation warning
was added. The deprecation warning is misleading since it has the wrong
date and references a `scan` command which does not exist. This will
remove the `scan` command from the documentation and update the
deprecation warning.

Reference-to: https://github.com/pypa/pipenv/issues/6396