Fix editable installation from relative paths on Windows
When running `pipenv install -e .\..\mypackage` the pip requirement
line `-e .\..\mypackage` is written to a temporary requirements file
which is passed to pip to install the package. pip fails to install
the package from the requirements file as it incorrectly interprets
the relative path, treating the slashes as escapes. To work around
this issue replace all back slashes (windows directory separators)
with forward slashes (POSIX directory separators) in pip requirements
for editable package installs.