Validate Python version when using --system flag
Fixes #6403

Previously, 'pipenv install --system --deploy' would silently proceed
even if the system Python version didn't match the python_version
specified in Pipfile/Pipfile.lock. This was because the version check
was only performed when creating a new virtualenv, not when using
--system.

This moves the Python version check outside of the virtualenv creation
block so it runs for both virtualenv and --system installations.
When --deploy is used with an incompatible Python version, a
DeployException is now properly raised.