Fix VCS dependencies not reinstalling when ref/commit changes (#6466)
The is_satisfied() method was returning True for editable VCS dependencies
as long as the source directory existed, without checking if the commit
hash matched the lockfile. This caused 'pipenv update' to update the lockfile
with the new commit but leave the old code installed in the virtualenv.

The fix moves the VCS check to the top of is_satisfied() to always return
False for VCS dependencies, forcing pip to reinstall and check out the
correct commit.

Fixes #5791