Add PIPENV_PROJECT_DIR environment variable (#6497)
Adds a new PIPENV_PROJECT_DIR environment variable that is automatically
set when running 'pipenv run' or 'pipenv shell'. This variable contains
the path to the project root directory (where the Pipfile is located).

This allows scripts defined in the [scripts] section of Pipfile to
reference project-relative paths regardless of the current working
directory:

  [scripts]
  test = "pytest $PIPENV_PROJECT_DIR/tests"
  mypy = "mypy --config-file=$PIPENV_PROJECT_DIR/mypy.ini src"

This is particularly useful when:
- Running pipenv commands from subdirectories
- Scripts need to reference config files in the project root
- You have a src layout and need consistent paths

Also documents the runtime environment variables (PIPENV_ACTIVE,
PIPENV_PROJECT_DIR, VIRTUAL_ENV) in the configuration docs.

Fixes #2241